Used long for key indexes

+++ KeyStore.c
@@ -10,7 +10,7 @@ extern "C"
 
 
 static
-void reset_element_key(key_store_t *key_store, unsigned int index)
+void reset_element_key(key_store_t *key_store, unsigned long index)
 {
     memset(key_store->element_store[index].key.name, 0, KEY_NAME_SIZE);
     memset(key_store->element_store[index].key.data, 0, KEY_DATA_SIZE);
@@ -18,7 +18,7 @@ void reset_element_key(key_store_t *key_store, unsigned int index)
 
 
 static
-void delete_element(key_store_t *key_store, unsigned int index)
+void delete_element(key_store_t *key_store, unsigned long index)
 {
     if (!key_store->element_store[index].admin.is_free)
     {
@@ -33,7 +33,7 @@ void delete_element(key_store_t *key_store, unsigned int index)
 
 
 static
-unsigned int find_element(key_store_t const *key_store, unsigned int max, const unsigned int app_id, const char name [KEY_NAME_SIZE])
+unsigned long find_element(key_store_t const *key_store, unsigned long max, const unsigned int app_id, const char name [KEY_NAME_SIZE])
 {
     if (name == NULL)
     {
@@ -45,7 +45,7 @@ unsigned int find_element(key_store_t const *key_store, unsigned int max, const 
         max = key_store->max_elements;
     }
 
-    for (unsigned int k = 0; k < max; k++)
+    for (unsigned long k = 0; k < max; k++)
     {
         if (!key_store->element_store[k].admin.is_free)
         {
@@ -61,9 +61,9 @@ unsigned int find_element(key_store_t const *key_store, unsigned int max, const 
 }
 
 static
-unsigned int find_free_element(key_store_t const *key_store)
+unsigned long find_free_element(key_store_t const *key_store)
 {
-    for (unsigned int k = 0; k < key_store->max_elements; k++)
+    for (unsigned long k = 0; k < key_store->max_elements; k++)
     {
         if (key_store->element_store[k].admin.is_free)
         {
@@ -76,13 +76,13 @@ unsigned int find_free_element(key_store_t const *key_store)
 
 void key_store_init(
     key_store_t *key_store,
-    unsigned max_elements,
+    unsigned long max_elements,
     element_record_t *element_store)
 {
     key_store->max_elements = max_elements;
     key_store->element_store = element_store;
     key_store->free_slots = key_store->max_elements;
-    for (unsigned int k = 0; k < key_store->max_elements; k++)
+    for (unsigned long k = 0; k < key_store->max_elements; k++)
     {
         key_store->element_store[k].admin.is_free = 1;
         key_store->element_store[k].admin.app_id = 0;
@@ -94,8 +94,8 @@ unsigned int key_store_init_with_read_only_keys(
     key_store_t *key_store,
     unsigned int const *app_ids,
     key_record_t const *keys,
-    unsigned int nr_keys,
-    unsigned max_elements,
+    unsigned long nr_keys,
+    unsigned long max_elements,
     element_record_t *element_store)
 {
     unsigned int result = 0;
@@ -109,7 +109,7 @@ unsigned int key_store_init_with_read_only_keys(
         result = -1;
     }
 
-    for (unsigned int k = 0; k < nr_keys; k++)
+    for (unsigned long k = 0; k < nr_keys; k++)
     {
         // if there is a duplicate, return an empty keystore
         if (find_element(key_store, k, app_ids[k], keys[k].name) < k)
@@ -126,7 +126,7 @@ unsigned int key_store_init_with_read_only_keys(
         memcpy(key_store->element_store[k].key.data, keys[k].data, KEY_DATA_SIZE);
     }
 
-    for (unsigned int k = nr_keys; k < key_store->max_elements; k++)
+    for (unsigned long k = nr_keys; k < key_store->max_elements; k++)
     {
         key_store->element_store[k].admin.is_free = 1;
         key_store->element_store[k].admin.app_id = 0;
@@ -141,7 +141,7 @@ unsigned int key_store_init_with_read_only_keys(
 
 void key_store_wipe(key_store_t *key_store)
 {
-    for (unsigned int k = 0; k < key_store->max_elements; k++)
+    for (unsigned long k = 0; k < key_store->max_elements; k++)
     {
         if (!key_store->element_store[k].admin.is_free &&
             !key_store->element_store[k].key.read_only)
@@ -231,7 +231,7 @@ key_store_result_t key_store_get(key_store_t const *key_store, unsigned int app_
     return result;
 }
 
-unsigned int key_store_get_by_index(key_store_t const *key_store, unsigned int app_id, unsigned int index, key_record_t *key)
+unsigned int key_store_get_by_index(key_store_t const *key_store, unsigned int app_id, unsigned long index, key_record_t *key)
 {
     if (key == NULL)
     {
@@ -277,7 +277,7 @@ unsigned int key_store_delete(key_store_t *key_store, unsigned int app_id, const
         return -1;
     }
 
-    unsigned int element_index = find_element(key_store, key_store->max_elements, app_id, name);
+    unsigned long element_index = find_element(key_store, key_store->max_elements, app_id, name);
 
     if (key_store->max_elements == element_index)
     {
