+++ KeyStore.h
@@ -26,29 +26,29 @@ typedef struct int_element_record {
 
 
 typedef struct int_key_store {
-    unsigned free_slots;
-    unsigned max_elements;
+    unsigned long free_slots;
+    unsigned long max_elements;
     element_record_t *element_store;
 } key_store_t;
 
 typedef struct int_key_store_result {
     signed error;
-    unsigned index;
+    unsigned long index;
 } key_store_result_t;
 
 void key_store_init(
     key_store_t *key_store,
-    unsigned max_elements,
+    unsigned long max_elements,
     element_record_t *element_store);
 unsigned int key_store_init_with_read_only_keys(
     key_store_t *key_store,
     unsigned int const *app_ids,
     key_record_t const *keys,
-    unsigned int nr_keys,
-    unsigned max_elements,
+    unsigned long nr_keys,
+    unsigned long max_elements,
     element_record_t *element_store);
 void key_store_wipe(key_store_t *key_store);
 key_store_result_t key_store_add(key_store_t *key_store, unsigned int app_id, key_record_t const *key);
 key_store_result_t key_store_get(key_store_t const *key_store, unsigned int app_id, const char name [KEY_NAME_SIZE], key_record_t *key);
-unsigned int key_store_get_by_index(key_store_t const *key_store, unsigned int app_id, unsigned int index, key_record_t *key);
+unsigned int key_store_get_by_index(key_store_t const *key_store, unsigned int app_id, unsigned long index, key_record_t *key);
 unsigned int key_store_delete(key_store_t *key_store, unsigned int app_id, const char name [KEY_NAME_SIZE]);
